requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-Admin
Flask-Swagger
python-dotenv
psycopg2-binary
